# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1"
    },
    "paths": {
...
        "/api/admin-panel/users": {
            "get": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Returns datas of user - > with filters and search option and sort options",
                "description": "Returns datas of user - > with filters and search option and sort options",
                "operationId": "App\\Http\\Controllers\\Dashboard\\UserController::index",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Your can provide filter option to all incomming datas in: customer,listener",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "customer"
                    },
                    {
                        "name": "searchName",
                        "in": "path",
                        "description": "U can provide search name ,that will be filtered and bring back datas like this name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Mike"
                    },
                    {
                        "name": "searchPhone",
                        "in": "path",
                        "description": "U can provide search phone ,that will be filtered and bring back datas like this phone.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "25252525"
                    },
                    {
                        "name": "searchEmail",
                        "in": "path",
                        "description": "U can provide search email ,that will be filtered and bring back datas like this email.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test@test.lv"
                    },
                    {
                        "name": "order",
                        "in": "path",
                        "description": "U can provide order/sort type,with all datas will be sorted and returned back. in:type | full_name | phone_number | email | rating | created_at | balance | date_of_birth | succesfull_calls | total_time",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "rating"
                    },
                    {
                        "name": "order_type",
                        "in": "path",
                        "description": "Your can provide order type -> in: desc,asc ( WORKS WHEN ORDER IS SPECIFIED)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": [
                                                {
                                                    "id": 63,
                                                    "full_name": null,
                                                    "photo": null,
                                                    "email": null,
                                                    "date_of_birth": "06.09.2021",
                                                    "phone_number": "+3712608379423",
                                                    "gender": null,
                                                    "type": "listener",
                                                    "rating": null,
                                                    "status": "offline",
                                                    "bio": null,
                                                    "email_verified_at": null,
                                                    "device_token_customer": null,
                                                    "device_token_listener": null,
                                                    "remember_token": null,
                                                    "created_at": "2021-08-11T16:33:32.000000Z",
                                                    "updated_at": "2021-08-11T16:34:18.000000Z",
                                                    "balance": 30,
                                                    "profile_url": null,
                                                    "language": null,
                                                    "stripe_id": null,
                                                    "last_activity_date": null,
                                                    "consent_to_agreement": 0,
                                                    "succesfull_calls": 1,
                                                    "total_time": 50
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}